+++ addon/helpers/get-formatted.js
@@ -29,8 +29,7 @@ export default Ember.Helper.extend({
         let dateFormat = Ember.get(hash, 'dateFormat');
         return dateFormat ? momentValue.format(dateFormat) : momentValue.format();
       case 'boolean':
-        return value ? this.get('i18n').t('components.object-list-view-cell.boolean-true-caption')
-          : this.get('i18n').t('components.object-list-view-cell.boolean-false-caption');
+        return Ember.String.htmlSafe(`<div class='ui checkbox disabled'><input type='checkbox' class='hidden' ${value ? 'checked' : ''}><label></label></div>`);
       default:
         return value;
     }